[Collaborative project] To-Do List Application (GUI/UI) - Java & Swing
Accomplished Leadership in overseeing a team of six individuals as measured by Collaborative Project Delivery, by spearheading the development of a To-Do List GUI Application using Java Swing. Organized and guided a team through the process of Designing, Implementing, and Validating the application's functionalities, resulting in a Robust Task Management System with intuitive user interfaces and seamless task handling.